Web content gets stuck in an inactive state (no cursor updates or text insertion caret) when activating a tab with a thumbnail visible https://bugs.webkit.org/show_bug.cgi?id=214962 <rdar://problem/65670984>
Reviewed by Wenson Hsieh.

Source/WebCore:

New API test: WebKit.WKThumbnailViewResetsViewStateWhenUnparented

* testing/Internals.cpp:
(WebCore::Internals::isPageActive const):
* testing/Internals.h:
* testing/Internals.idl:
Add a "view window is active" getter.

Source/WebKit:

* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::setThumbnailView):
When WKThumbnailView is unparented, invalidate all activity state bits.
We do this because many of the functions that compute activity state bits
use the WKThumbnailView's window while it is parented. When it is unparented,
and we switch back to using the WKWebView's window, we mustn't get stuck
with activity state bits from WKThumbnailView's window.

This was particularly problematic in the case of the WKThumbnailViews
used for tab hover previews, because that window is not key, so we'd get
stuck with the inactive-window state, resulting in a lack of cursor updates,
a hidden text caret, the wrong selection color, etc.

* UIProcess/mac/PageClientImplMac.mm:
(WebKit::PageClientImpl::isViewWindowActive):
While writing the API test for this, I noticed that isViewWindowActive
returns YES if there is no key window AND WKWebView is unparented
([NSApp keyWindow] == activeWindow() == nil).
This seems completely insane (how can the window be active if it doesn't
exist?), and breaks the test, so ensure that we only say YES if we actually
have a window.

Tools:

* TestWebKitAPI/Tests/WebKit/WKThumbnailView.mm:
(TestWebKitAPI::TEST):
Add a test ensuring that we go back to the active-window state after
unparenting a WKThumbnailView that was installed in an inactive window.

(-[WKThumbnailViewDelegate webView:didFinishNavigation:]): Deleted.
* TestWebKitAPI/mac/OffscreenWindow.h:
* TestWebKitAPI/mac/OffscreenWindow.mm:
(-[OffscreenWindow initWithSize:]):
(-[OffscreenWindow initWithSize:isKeyWindow:]):
(-[OffscreenWindow isKeyWindow]):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@265091 268f45cc-cd09-0410-ab3c-d52691b4dbfc